# apps/banking-api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Intergalactic Banking API
  version: 2.0.0
  description: A deterministic, run-scoped Banking API for comparing direct agent integrations with Postman
    Fabric Gateway. Monetary values are non-negative safe integers in minor units; all fictional currencies
    currently use exponent 0.
  contact:
    name: Intergalactic Banking API team
    url: https://www.postman.com/
  license:
    name: ISC
servers:
  - url: http://localhost:3000
    description: Local demo server
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/transactions:
    get:
      tags:
        - Transactions
      summary: List visible transactions
      description: List visible transactions in the selected run.
      operationId: listTransactions
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: fromAccountId
          in: query
          required: false
          description: Optional fromAccountId filter.
          schema:
            type: string
        - name: toAccountId
          in: query
          required: false
          description: Optional toAccountId filter.
          schema:
            type: string
        - name: createdAt
          in: query
          required: false
          description: Optional createdAt filter.
          schema:
            type: string
      responses:
        "200":
          description: List visible transactions succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags:
        - Transactions
      summary: Post an atomic transfer or deposit
      description: Post an atomic transfer or deposit in the selected run.
      operationId: createTransaction
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Post an atomic transfer or deposit succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
  /api/v1/transactions/{transactionId}:
    get:
      tags:
        - Transactions
      summary: Get a visible transaction
      description: Get a visible transaction in the selected run.
      operationId: getTransaction
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - name: transactionId
          in: path
          required: true
          description: Run-scoped transactionId value.
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Get a visible transaction succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: A credential resolving to a CUSTOMER or ADMIN principal.
  parameters:
    DemoRunId:
      name: X-Demo-Run-Id
      in: header
      required: false
      description: Isolates state for a comparison lane; defaults to default.
      schema:
        type: string
        pattern: ^[A-Za-z0-9_-]{1,64}$
        default: default
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Caller-provided correlation identifier; generated deterministically when omitted.
      schema:
        type: string
        minLength: 1
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Required for mutations. Exact replays return the original sanitized response; conflicting or
        concurrent duplicates return 409.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    Limit:
      name: limit
      in: query
      required: false
      description: Page size.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque collection-scoped continuation cursor.
      schema:
        type: string
  headers:
...
    Retry-After:
      description: Seconds until another request should be attempted.
      schema:
        type: integer
        minimum: 1
    Idempotency-Replayed:
      description: True when the response was replayed.
      schema:
        type: string
        const: "true"
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          required:
            - name
            - message
          properties:
            name:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
    Page:
      type: object
      additionalProperties: false
      required:
        - limit
        - nextCursor
        - hasMore
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 100
        nextCursor:
          type:
            - string
            - "null"
        hasMore:
          type: boolean
...
    Transaction:
      type: object
      additionalProperties: false
      required:
        - transactionId
        - createdAt
        - amount
        - currency
        - fromAccountId
        - toAccountId
        - status
        - description
        - beneficiaryName
        - channel
        - deviceTrusted
        - location
      properties:
        transactionId:
          type: string
        createdAt:
          type: string
          format: date-time
        amount: *a6
        currency: *a7
        fromAccountId:
          type: string
        toAccountId:
          type: string
        status:
          type: string
          enum:
            - POSTED
        description:
          type:
            - string
            - "null"
        beneficiaryName:
          type:
            - string
            - "null"
        channel:
          type:
            - string
            - "null"
        deviceTrusted:
          type:
            - boolean
            - "null"
        location:
          type:
            - string
            - "null"
...
    TransactionResponse:
      type: object
      additionalProperties: false
      required:
        - transaction
      properties:
        transaction:
          $ref: "#/components/schemas/Transaction"
    TransactionList:
      type: object
      additionalProperties: false
      required:
        - transactions
        - page
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        page:
          $ref: "#/components/schemas/Page"
...
    CreateTransactionRequest:
      type: object
      additionalProperties: false
      required:
        - fromAccountId
        - toAccountId
        - amount
        - currency
      properties:
        fromAccountId:
          type: string
          minLength: 1
        toAccountId:
          type: string
          minLength: 1
        amount: &a9
          type: integer
          minimum: 1
          maximum: 9007199254740991
        currency: *a8
        description:
          type: string
        beneficiaryName:
          type: string
        channel:
          type: string
          enum:
            - MOBILE_APP
            - WEB
            - API
            - BRANCH
        deviceTrusted:
          type: boolean
        location:
          type: string
...
  responses:
    BadRequest:
      description: The request is malformed or violates a schema.
      headers: &a10
        X-Demo-Run-Id: *a1
        X-Request-Id: *a2
        X-RateLimit-Limit: *a3
        X-RateLimit-Remaining: *a4
        X-RateLimit-Reset: *a5
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: A valid API key is required.
      headers: *a10
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The principal lacks permission.
      headers: *a10
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The resource does not exist or is not visible.
      headers: *a10
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: A state transition, uniqueness rule, or idempotency reservation conflicted.
      headers: *a10
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: The run-and-credential rate limit was exceeded.
      headers:
        X-Demo-Run-Id: *a1
        X-Request-Id: *a2
        X-RateLimit-Limit: *a3
        X-RateLimit-Remaining: *a4
        X-RateLimit-Reset: *a5
        Retry-After:
          $ref: "#/components/headers/Retry-After"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: A bounded run or idempotency store reached capacity.
      headers: *a10
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"